Fix Dequantize.inverse to return Long tensor matching original dtype
torch.floor() returns a float; cast back to long so the inverse
properly round-trips integer num_points data.

+++ pointcountfm/preprocessing.py
@@ -159,7 +159,7 @@ class Dequantize(Transformation):
         return x.float() + torch.rand_like(x.float())
 
     def inverse(self, x: torch.Tensor) -> torch.Tensor:
-        return torch.floor(x)
+        return torch.floor(x).long()
 
 
 def compose(transformation: list[list[str | dict | list | None]] | None) -> Sequence:
